last round of shuffling, I think
