+++ src/CribblyBackend.DataAccess/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,7 @@
-using System;
+using CribblyBackend.Core.Games.Repositories;
+using CribblyBackend.Core.Players.Repositories;
+using CribblyBackend.Core.Teams.Repositories;
+using CribblyBackend.Core.Tournaments.Repositories;
 using CribblyBackend.DataAccess.Games.Repositories;
 using CribblyBackend.DataAccess.Players.Repositories;
 using CribblyBackend.DataAccess.Teams.Repositories;
